+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0]
+
+### Added
+
+- Shipments can now be sent as physical mail (Fysisk Post) or as Digital Post with physical fallback.
+
+### Fixed
+
+- Added error handling on syntax errors in templates.
+- Bumped nginx max file size to 10mb.
+
 ## [0.1.0]
 
 ### Added
@@ -30,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/itk-dev-rpa/OpenPostbud/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/itk-dev-rpa/OpenPostbud/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/itk-dev-rpa/OpenPostbud/releases/tag/0.2.0
 [0.1.0]: https://github.com/itk-dev-rpa/OpenPostbud/releases/tag/0.1.0
 [0.0.1]: https://github.com/itk-dev-rpa/OpenPostbud/releases/tag/0.0.1